+++ assertr/src/assertions/matcher.rs
@@ -114,7 +114,7 @@ mod tests {
                 it.is_equal_to(1).is_less_than(2);
             }));
 
-            assert_that!(subject.state.number_of_assertions.borrow().0).is_equal_to(1);
+            assert_that!(subject.state.records.assertion_count()).is_equal_to(1);
         }
 
         #[test]